e2e tests: stop renaming .zip files into .xpi
Selenium became able to load .zip extensions in Firefox a couple of
months ago:
https://github.com/SeleniumHQ/selenium/commit/4ea31198a7808ae7713be337d21f168156e26aba
