+++ package.json
@@ -59,7 +59,7 @@
     "make-zip": "web-ext build --source-dir build --overwrite-dest",
     "pretty": "bash scripts/pretty.sh",
     "run": "web-ext run -s build/ -u 'txti.es'",
-    "test": "yarn run build && rm -rf web-ext-artifacts/* && web-ext build --source-dir ./build --overwrite-dest && mv web-ext-artifacts/*.zip web-ext-artifacts/tridactyl.xpi && jest --silent",
+    "test": "yarn run build && web-ext build --source-dir ./build --overwrite-dest && jest --silent",
     "update-buildsystem": "rm -rf src/node_modules; yarn run clean",
     "watch": "echo 'watch is broken, use build instead'; exit 0;"
   },
